Python/Steganography_Hide_Text_In_Image/requirements.txt
Pillow
stegano
